ai_engine/requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
requests
PyPDF2
python-dotenv